chore(size): give serverDataModel.ts and columnar-parser.ts real headroom
Both files landed this PR sitting exactly at their check-module-size.mjs
budget (630/630 and 1189/1189) — the same "growth from an unrelated PR
tips it over" pattern that broke main via CommandPalette.tsx earlier
(fixed there by extraction in #3958). Extract a coherent seam from each
and lower its allowlist entry to leave real slack:

- apps/viewer/src/utils/serverDataModel.ts: pulled the whole "Spatial
  Hierarchy Building" section (ServerSpatialNode, buildSpatialNodeTree,
  buildSpatialHierarchy) into serverSpatialHierarchy.ts. 630 -> 444 lines;
  budget lowered to 470 (26 lines headroom).
- packages/parser/src/columnar-parser.ts: pulled the root-attribute
  resolution helpers (getEntityRefFromStore, extractEntityAttributesOnDemand,
  extractAllEntityAttributes, getRawNamedAttributes, getRootAttrIndices,
  extractRootAttributesFromEntity, pickLongName) into
  columnar-parser-root-attributes.ts, re-exported for the existing public
  surface. 1189 -> 946 lines; budget lowered to 1000 (54 lines headroom).
  Kept clear of the re-export block near the end of the file that #3951
  also touches.

Also un-compacted two lines this PR's own diff had squeezed to fit the old
budget: the classification regroup loop in serverDataModel.ts (was a single
`for (...) resolvedClassifications.set(...)` line) and the
`resolvedClassifications` field's doc comment in columnar-parser.ts (was
sharing a line with the field declaration).
